Refactor OpenAI helpers into modular utilities
