+++ src/config/openaiPrompts.ts
@@ -1,0 +1,1 @@
+export const STRICT_ASSISTANT_PROMPT = 'You are a precise and safe code assistant.';
